move openscadtestframework to local module
