+++ tests/openscadtestframework/__init__.py
@@ -1,0 +1,3 @@
+# __init__.py
+from .openscadtestframework import ScadTestCase, Module, ModuleTest
+from .basic_modules import Cube, Square
